use object-position for layout of images
